### 2026-06-07 #10662 @editorialbot
[PRE REVIEW]: yabplot: yet another brain plot for unified neuroimaging visualization in Python
**Submitting author:** @teanijarv (<a href="http://orcid.org/0000-0002-3650-4230">Toomas Erik Anijärv</a>)
**Repository:** https://github.com/teanijarv/yabplot
**Branch with paper.md** (empty if default branch):
**Version:** 0.5.1
