### 2026-06-24 a714c11
Fix default argument behavior of cmap if data not provided
