### 2026-06-25 90552b2
add lut support in plot_vertexwise()
