### 2026-07-21 ac21ac6
Update paper.md
