+++ paper/paper.md
@@ -23,13 +23,13 @@ Neuroimaging analyses often produce results in several anatomical representation
 
 `yabplot` (yet another brain plot) is an open-source Python package for creating three-dimensional neuroimaging visualizations across these representations. It provides a unified interface for plotting cortical parcellations, vertex-wise cortical data, voxel-wise volumetric data, subcortical structures, white-matter bundles, tractometry results, and connectome graphs. The package is intended for researchers who want to create publication-ready figures directly from Python scripts or Jupyter notebooks, without switching between multiple specialized graphical applications. Supported inputs depend on the plotting function and include NIfTI images, surface files, tractography files, arrays, dictionaries, tabular data, and matrices. These inputs are mapped to standard or user-defined anatomical resources.
 
-The package builds on established scientific Python libraries. It uses `nibabel` for neuroimaging file input/output [@Brett:2024], `numpy` for numerical array operations [@Harris:2020], `pandas` for tabular and labelled data inputs [@McKinney:2010], `scipy` for interpolation, filtering, and sparse matrix operations [@Virtanen:2020], `pooch` for resource retrieval and caching [@Uieda:2020], `scikit-image` for extracting surface meshes from volumetric masks [@vanDerWalt:2014], `pyvista` for three-dimensional mesh representation, scene construction, and rendering [@Sullivan:2019], `trame` for interactive browser-based visualization [@Jourdain:2025], and `matplotlib` for static figure composition and customization [@Hunter:2007]. By combining these tools behind a compact plotting API, `yabplot` handles data mapping, mesh creation, camera placement, lighting, and rendering while retaining compatibility with standard Python plotting workflows.
+The package builds on established scientific Python libraries. It uses `nibabel` for neuroimaging file I/O [@Brett:2024], `numpy` for array operations [@Harris:2020], `pandas` for tabular and labelled data inputs [@McKinney:2010], `scipy` for interpolation and sparse matrix operations [@Virtanen:2020], `pooch` for resource retrieval and caching [@Uieda:2020], `scikit-image` for surface-mesh extraction from volumetric masks [@vanDerWalt:2014], `pyvista` for 3D mesh representation and rendering [@Sullivan:2019], `trame` for interactive browser-based visualization [@Jourdain:2025], and `matplotlib` for static figure composition [@Hunter:2007]. By combining these tools behind a compact plotting API, `yabplot` handles data mapping, mesh creation, camera placement, lighting, and rendering while retaining compatibility with standard Python plotting workflows.
 
 ![Overview of `yabplot` functionality. The package provides utilities for resource access, atlas construction, volume projection, and mesh handling, together with high-level plotting functions for cortical parcellations, vertex-wise surface maps, subcortical structures, voxel-wise volumes, white matter tracts, and connectome graphs.](figures/overview_joss.pdf)
 
 # Statement of need
 
-Researchers working with neuroimaging data often need to communicate results across several anatomical representations, including cortical regions, vertex-wise maps, subcortical structures, voxel-wise maps, white matter tracts, and connectivity matrices [@Chopra:2023]. The target users of `yabplot` are neuroimaging researchers and students who want to generate consistent 3D figures directly in Python-based data analysis workflows.
+Researchers working with neuroimaging data often need to communicate results across several anatomical representations [@Chopra:2023]. The target users of `yabplot` are neuroimaging researchers and students who want to generate consistent 3D figures directly in Python-based data analysis workflows.
 
 Producing such figures can be difficult when each representation requires a different application or plotting library. Switching between tools can require data conversion, manual screenshot-based workflows, and repeated adjustment of lighting, camera angles, colormaps, and output formatting. `yabplot` addresses this need by providing a unified, scriptable interface for common 3D neuroimaging visualizations, with consistent styling and output handling across modalities.
 
