### 2026-07-24 caeed31
fix colormap-object crash and midline filename collision in plot_subcortical
