### 2026-07-24 09a960b
make per-region export optional and replace shuffle_colors with atlas_color_seed
