+++ yabplot/plotting.py
@@ -1,6 +1,7 @@
 import os
 import gc
 import re
+import random
 import warnings
 import numpy as np
 import pandas as pd
@@ -127,7 +128,7 @@ def get_base_name(name):
     for suffix in ['_l', '_r', '-lh', '-rh']:
         if n.endswith(suffix):
             return name[:len(name)-len(suffix)]
-    for prefix in ['l-', 'r-']:
+    for prefix in ['l-', 'r-', 'l_', 'r_']:
         if n.startswith(prefix):
             return name[len(prefix):]
     for word in ['left_', 'right_', 'left-', 'right-']:
@@ -336,8 +337,8 @@ def plot_subcortical(data=None, atlas=None, custom_atlas_path=None, ax=None, cba
                      nan_alpha=1.0, style='default', bmesh='midthickness',
                      bmesh_alpha=0.15, bmesh_color='lightgray', zoom=1.2, display_type='matplotlib',
                      export_path=None, custom_atlas_proc=dict(smooth_i=15, smooth_f=0.6),
-                     shuffle_colors=False, plot_regions_separately=False,
-                     hemisphere_colors=False):
+                     atlas_color_seed=42, plot_regions_separately=False,
+                     shared_hemisphere_colors=False):
     """
     Visualize data on the subcortical structures using a specified atlas.
 
@@ -398,18 +399,23 @@ def plot_subcortical(data=None, atlas=None, custom_atlas_path=None, ax=None, cba
         Parameters for processing custom GIfTI files.
         Keys: 'smooth_i' (iterations) and 'smooth_f' (relaxation factor).
         Default is {'smooth_i': 15, 'smooth_f': 0.6}.
-    shuffle_colors : bool, optional
-        If True, shuffles the color assignments for categorical coloring to help
-        distinguish adjacent regions with similar colormap neighbors. Default is False.
+    atlas_color_seed : int, optional
+        Seed controlling the categorical color arrangement in atlas (no-data) mode.
+        With the default palette it seeds the distinct-color generator; with a
+        colormap it seeds a shuffle so gradient-adjacent regions get contrasting
+        colors. Change it to get a different, reproducible arrangement. Ignored when
+        `data` is given. Default is 42.
     plot_regions_separately : bool, optional
-        If True, saves a separate PNG file for each region + view combination instead
-        of a single combined figure. Each file is named ``{region}_{view_face}.png``
-        and uses a transparent background (useful for SVG tracing pipelines).
-        ``export_path`` is treated as an output directory. Default is False.
-    hemisphere_colors : bool, optional
-        If True, bilateral region pairs (e.g. ``putamen_l`` / ``putamen_r``) share the
-        same color in atlas (no-data) mode. If False (default), every region is assigned
-        its own independent color.
+        If True, renders each region + view combination on its own instead of a single
+        combined figure. If ``export_path`` is set, one transparent-background PNG per
+        region is saved there (named ``{region}_{view_face}.png``, useful for SVG
+        tracing pipelines) and None is returned; otherwise the per-region axes are
+        displayed and returned as a list. Default is False.
+    shared_hemisphere_colors : bool, optional
+        If True, bilateral region pairs (e.g. ``putamen_l`` / ``putamen_r``) are drawn
+        in the same color, so each structure reads as one color across both hemispheres.
+        Only applies in atlas (no-data) mode; ignored when `data` is given. If False
+        (default), every region is assigned its own independent color.
 
     Returns
     -------
@@ -453,31 +459,24 @@ def plot_subcortical(data=None, atlas=None, custom_atlas_path=None, ax=None, cba
         vmax = vminmax[1] if vminmax[1] is not None else (max(valid_vals) if valid_vals else 1)
         c_vlim = [vmin, vmax]
[…]
