+++ paper/paper.bib
@@ -170,9 +170,10 @@
 @article{Bryant:2026,
   title     = {Subcortex visualization: A toolbox for custom data visualization in the subcortex and cerebellum},
   author    = {Bryant, Annie G.},
-  journal   = {bioRxiv},
+  journal   = {Aperture Neuro},
+  volume    = {6},
   year      = {2026},
-  publisher = {Cold Spring Harbor Laboratory},
-  doi       = {10.64898/2026.01.23.699785},
-  url       = {https://doi.org/10.64898/2026.01.23.699785}
+  publisher = {Organization for Human Brain Mapping},
+  doi       = {10.52294/001c.165884},
+  url       = {https://doi.org/10.52294/001c.165884}
 }
