### 2026-08-20 #11157 @teanijarv
Hi all! Just letting you know that I just updated one DOI which previously was a pre-print, but now had been published in Aperture Neuro journal. Looking forward to your reviews.
